requirements.txt
crewai
langchain_core
google-generativeai
langchain